requirements.txt
# UAE Talent Intelligence — Python dependencies
# Install with: pip install -r requirements.txt

# Core ETL
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3

# Data transformation
numpy==1.26.4
scikit-learn==1.5.0         # salary imputation model
rapidfuzz==3.9.3            # fuzzy company name matching

# Utilities
tqdm==4.66.4                # progress bars
colorama==0.4.6             # coloured terminal output
tabulate==0.9.0             # pretty-print summary tables
tenacity==8.3.0             # retry logic for API calls